Slightly rework docs release process
